Move all npointfunction observables to new folder.
