+++ templates/observables.cpp.in
@@ -23,8 +23,8 @@
 #include "@ModelName@_edm.hpp"
 #include "@ModelName@_l_to_lgamma.hpp"
 #include "@ModelName@_b_to_s_gamma.hpp"
-#include "npointfunctions/@ModelName@_l_to_l_conversion.hpp"
-#include "npointfunctions/l_to_l_conversion/settings.hpp"
+#include "npointfunctions/@ModelName@_ltolconversion.hpp"
+#include "npointfunctions/ltolconversion/settings.hpp"
 #include "npointfunctions/@ModelName@_lto3l.hpp"
 #include "config.h"
 #include "eigen_utils.hpp"
